Add files via upload
Summary of the compounds screened for the ATP binding site.  The compounds were designed by Joe Eyermann and Rebecca Stevenson as potential dual inhibitors for MurD and MurE. The sheet is colour coded according to the stage in which the experiment failed. Red - failed during soaking/fishing; Orange - data collection. Cells coloured green indicates data samples that was of low resolution.